requirements.txt
mlflow==2.14.1
numpy
pandas
pyyaml
tqdm
torch
torchvision
scikit-learn
matplotlib